+++ src/main/java/by/vyun/bgb/repository/RatingRepo.java
@@ -4,6 +4,6 @@ import by.vyun.bgb.entity.Rating;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RatingRepo extends JpaRepository<Rating, Integer> {
-    Rating findGameRatingByUserIdAndGameId(int userId, int gameId);
+    Rating findGameRatingByUserIdAndGameId(int userId, long gameId);
 
 }
